scraper/requirements.txt
requests
kafka-python
feedparser
beautifulsoup4